put tests into one binary
This refactor puts tests into a single binary that runs with
multiprocess; both a simple runner that just runs a collection and a
dylib that can be run with test versions of the main yachtrock dylib and
a test version of yr_runtests that links it. Fixed a few bugs as a
result.

resolves #5